Fix compilation of non playground shaders
